+++ CHANGELOG.md
@@ -15,7 +15,7 @@
   - Include reffered vars in completion. #1440
   - Allow rename alias of function calls, renaming properly both function usages and alias definition.
   - Add support for keyword analysis in edn files, allowing find-references, document highlight and other lsp features to keywords.
-  - Navigation for quoted symbols like `'clojure.core/inc` and symbols mentioned in `.edn` files #1446
+  - find-definition and references for quoted symbols like `'clojure.core/inc` and symbols mentioned in `.edn` files. #1446
 
 - API/CLI
   - Make diagnostics command print as json or edn. #1419